+++ tests/testthat/test-lof.R
@@ -29,14 +29,14 @@ expect_equal(lof, lof_d)
 ## missing values, but distances are fine
 x_na <- x
 x_na[c(1,3,5), 1] <- NA
-expect_error(lof(x_na), regexp = "NA")
-res_d1 <- lof(x_na, search = "dist")
-res_d2 <- lof(dist(x_na))
+expect_error(lof(x_na, k=4), regexp = "NA")
+res_d1 <- lof(x_na, k =4, search = "dist")
+res_d2 <- lof(dist(x_na), k = 4)
 expect_equal(res_d1, res_d2)
 
 x_na[c(1,3,5), 2] <- NA
-expect_error(lof(x_na), regexp = "NA")
-expect_error(lof(x_na, search = "dist"),
+expect_error(lof(x_na, k=4), regexp = "NA")
+expect_error(lof(x_na, k=4, search = "dist"),
   regexp = "NA")
-expect_error(lof(dist(x_na)), regexp = "NA")
+expect_error(lof(dist(x_na), k =4), regexp = "NA")