+++ R/LOF.R
@@ -17,9 +17,18 @@
 # with this program; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-lof <- function(x, k, ...) {
+lof <- function(x, minPts = 5, ...) {
 
+  ### parse extra parameters
   extra <- list(...)
+
+  # check for deprecated k
+  if(!is.null(extra[["k"]])) {
+    minPts <- extra[["k"]] + 1
+    extra[["k"]] <- NULL
+    warning("lof: k is now deprecated. use minPts = ", minPts, " instead .")
+  }
+
   args <- c("search", "bucketSize", "splitRule", "approx")
   m <- pmatch(names(extra), args)
   if(any(is.na(m))) stop("Unknown parameter: ",
@@ -32,46 +41,43 @@ lof <- function(x, k, ...) {
   splitRule <- .parse_splitRule(splitRule)
   bucketSize <- if(is.null(extra$bucketSize)) 10L else
     as.integer(extra$bucketSize)
-  approx <- if(is.null(extra$approx)) 0L else as.integer(extra$approx)
+  approx <- if(is.null(extra$approx)) 0 else as.double(extra$approx)
 
-  ### dist search
+  ### precompute distance matrix for dist search
   if(search == 3) {
     if(!inherits(x, "dist"))
       if(.matrixlike(x)) x <- dist(x)
       else stop("x needs to be a matrix to calculate distances")
   }
 
-  # get n
+  # get and check n
   if(inherits(x, "dist")) n <- attr(x, "Size")
   else n <- nrow(x)
-
   if(is.null(n)) stop("x needs to be a matrix or a dist object!")
-
-  if(k<1 || k>=n)
-    stop("k has to be larger than 1 and smaller than the number of points")
+  if(minPts<2 || minPts>n)
+    stop("minPts has to be at least 2 and not larger than the number of points")
 
 
   ### get LOF from a dist object
   if(inherits(x, "dist")) {
-
     if(any(is.na(x))) stop("NAs not allowed in dist for LOF!")
 
     # find k-NN distance, ids and distances
     x <- as.matrix(x)
     diag(x) <- Inf ### no self-matches
     o <- t(apply(x, 1, order, decreasing = FALSE))
-    k_dist <- x[cbind(o[,k], seq_len(n))]
+    k_dist <- x[cbind(o[,minPts-1], seq_len(n))]
     ids <- lapply(seq_len(n), FUN = function(i) which(x[i, ] <= k_dist[i]))
     dist <- lapply(seq_len(n), FUN = function(i) x[i, x[i, ] <= k_dist[i]])
 
-    ret <- list(dist = dist, ids = ids, k_dist = k_dist)
+    ret <- list(k_dist = k_dist, ids = ids, dist = dist)
 
   }else{
-
-    if(any(is.na(x))) stop("NAs not allowed for LOF using kdtree!")
     ### Use kd-tree
 
-    ret <- lof_kNN(as.matrix(x), as.integer(k),
+    if(any(is.na(x))) stop("NAs not allowed for LOF using kdtree!")
+
+    ret <- lof_kNN(as.matrix(x), as.integer(minPts),
       as.integer(search), as.integer(bucketSize),
       as.integer(splitRule), as.double(approx))
   }